all kinds of stuff before big type changes

+++ backend/routes/games/deleteGame.ts
@@ -1,4 +1,5 @@
 import { NextApiRequest, NextApiResponse } from 'next'
+import { Game } from '@backend/models'
 import { collections, throwError, verifyUser } from '@backend/utils'
 import compareObjectIds from '@backend/utils/compareObjectIds'
 import { objectIdSchema } from '@backend/validations/objectIdSchema'
@@ -8,7 +9,7 @@ const deleteGame = async (req: NextApiRequest, res: NextApiResponse) => {
   if (!userId) return throwError(res, 401, 'Unauthorized')
 
   const gameId = objectIdSchema.parse(req.query.id)
-  const game = await collections.games?.findOne({ _id: gameId })
+  const game = (await collections.games?.findOne({ _id: gameId })) as Game
 
   if (!game) {
     return throwError(res, 401, 'The game you are trying to delete could not be found')
@@ -16,6 +17,10 @@ const deleteGame = async (req: NextApiRequest, res: NextApiResponse) => {
 
   if (!compareObjectIds(userId, game.userId)) {
     return throwError(res, 401, 'You are not authorized to delete this game')
+  }
+
+  if (game.state === 'finished') {
+    return throwError(res, 400, 'Can only delete unfinished games')
   }
 
   const deletedGame = await collections.games?.deleteOne({ _id: gameId })
